move conftest from core to src
