+++ app.py
@@ -2,10 +2,7 @@
 import asyncio
 from argparse import ArgumentParser
 
-from src.core import ArgumentParserFacade
-from src.core.argument_default_value_from_config_provider import (
-    ArgumentDefaultValueFromConfigProvider,
-)
+from src.core import ArgumentDefaultValueFromConfigProvider, ArgumentParserFacade
 from src.protostar_cli import current_project, protostar_cli
 
 parser = ArgumentParserFacade(
