+++ src/core/reference_docs_generator_test.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from src.core.cli import CLI
 from src.core.command import Command
-from src.core.conftest import FooCommand
+from src.conftest import FooCommand
 from src.core.reference_docs_generator import ReferenceDocsGenerator
