+++ src/protostar_cli_test.py
@@ -1,0 +1,8 @@
+# from pytest_mock import MockerFixture
+# from .protostar_cli import ProtostarCLI, protostar_cli
+
+# def test_should_call_all_provided_commands(mocker: MockerFixture):
+#     cli = ProtostarCLI(version_manager=mocker.MagicMock(), commands)
+
+# # test_should_fail_due_to_old_git
+# # test_instance_matches_cli_reference_docs
